on rsync if not on server

+++ scripts/export_data.sh
@@ -7,6 +7,8 @@ export db_name=data_management
 export db_port=5432
 export db_user=developer
 
-rsync -rlpvhPit --delete datamanagement@$address:/media/tmc/nginx/81/uploads/ $DATA_MANAGEMENT_MEDIA_ROOT
+if [ $DATA_MANAGEMENT_MEDIA_ROOT != "/media/tmc/nginx/81/uploads/" ]; then
+  rsync -rlpvhPit --delete datamanagement@$address:/media/tmc/nginx/81/uploads/ $DATA_MANAGEMENT_MEDIA_ROOT
+fi
 
 source scripts/export_data_common.sh
